UI Shenaniagns, Code Cleanup
- Fix some issues with certain UI elements acting oddly on non-default themes
- Make some other UI issues a bit worse in some cases
- Fix a potential memory leak
- General code cleanup
- Localize a bunch more things

+++ CollectorsAnxiety/Data/Unlockables/Mount.cs
@@ -16,7 +16,7 @@ public class MountEntry : DataEntry<Mount> {
     public int NumberSeats => 1 + this.LuminaEntry.ExtraSeats;
     
     public override bool IsUnlocked() {
-        return CollectorsAnxietyPlugin.Instance.GameState.IsMountUnlocked((uint) this.Id);
+        return CollectorsAnxietyPlugin.Instance.GameState.IsMountUnlocked(this.Id);
     }
 
     public override bool IsValid() {
